Added tests to removeGameFromSite method
Added null checks to controller method

+++ src/main/java/com/petproject/boardgamefun/controller/GameController.java
@@ -113,6 +113,8 @@ public class GameController {
     @DeleteMapping("/remove")
     @PreAuthorize("hasRole('MODERATOR') or hasRole('ADMIN')")
     public ResponseEntity<String> removeGameFromSite(@RequestBody Game deleteGame) {
+        if (deleteGame.getId() == null)
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         gameRepository.delete(deleteGame);
 
         return new ResponseEntity<>(deleteGame.getTitle() + " удалена из базы данных", HttpStatus.OK);
